:sparkles: feature: 添加Education表的增删改查代码 :bug: bugfix: 修复部分增删改查业务bug :white_check_mark: test: 对绝大部分接口进行接口测试

+++ src/main/java/com/voyager/service/impl/ApplicationReviewServiceImpl.java
@@ -82,6 +82,6 @@ public class ApplicationReviewServiceImpl implements ApplicationReviewService {
     }
 
     public int deleteByJobId(int jobId) {
-         return applicationReviewMapper.deleteByJobId(jobId);
+         return applicationReviewMapper.deleteByJobId((long) jobId);
     }
 }
